Fixed sample and sample interval types

+++ include/adc.h
@@ -1,5 +1,5 @@
 #pragma once
 
 int adc_init(void);
-int adc_voltage_get(void);
-int adc_sample_interval_get(void);
+uint16_t adc_voltage_get(void);
+uint32_t adc_sample_interval_get(void);
